django/requirements.txt
Django
gunicorn
psycopg2-binary
django-allauth
martor
django-mptt